requirements.txt
fastapi
uvicorn
litellm
python-dotenv
httpx
rich
pandas
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http